perf: moved asset sorting into selectors
