+++ src/components/AssetSearch/AssetSearch.tsx
@@ -2,10 +2,9 @@ import { SearchIcon } from '@chakra-ui/icons'
 import type { BoxProps, InputProps } from '@chakra-ui/react'
 import { Box, Input, InputGroup, InputLeftElement, SlideFade } from '@chakra-ui/react'
 import type { Asset } from '@shapeshiftoss/asset-service'
-import type { AccountId, ChainId } from '@shapeshiftoss/caip'
+import type { ChainId } from '@shapeshiftoss/caip'
 import { debounce } from 'lodash'
 import intersection from 'lodash/intersection'
-import orderBy from 'lodash/orderBy'
 import uniq from 'lodash/uniq'
 import type { FC, FormEvent } from 'react'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
@@ -14,15 +13,8 @@ import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { useHistory } from 'react-router'
 import { Card } from 'components/Card/Card'
-import { bnOrZero } from 'lib/bignumber/bignumber'
 import { logger } from 'lib/logger'
-import {
-  selectChainIdsByMarketCap,
-  selectMarketDataSortedByMarketCap,
-  selectPortfolioFiatBalances,
-  selectPortfolioFiatBalancesByAccount,
-} from 'state/slices/selectors'
-import { useAppSelector } from 'state/store'
+import { selectChainIdsByMarketCap } from 'state/slices/selectors'
 
 import { AssetList } from './AssetList'
 import { ChainList } from './Chains/ChainList'
@@ -43,9 +35,7 @@ export enum AssetSearchOrder {
 export type AssetSearchProps = {
   assets: Asset[]
   onClick?: (asset: Asset) => void
-  sortOrder?: AssetSearchOrder
   disableUnsupported?: boolean
-  accountId?: AccountId
   assetListAsDropdown?: boolean
   hideZeroBalanceAmounts?: boolean
   formProps?: BoxProps
@@ -55,16 +45,13 @@ export const AssetSearch: FC<AssetSearchProps> = ({
   assets,
   onClick,
   disableUnsupported,
-  sortOrder = AssetSearchOrder.UserBalanceMarketCap,
-  accountId,
   assetListAsDropdown,
   hideZeroBalanceAmounts,
   formProps,
 }) => {
   const translate = useTranslate()
   const history = useHistory()
-  const portfolioFiatBalances = useSelector(selectPortfolioFiatBalances)
-  const portfolioFiatBalancesByAccount = useSelector(selectPortfolioFiatBalancesByAccount)
+
   const chainIdsByMarketCap = useSelector(selectChainIdsByMarketCap)
   const [activeChain, setActiveChain] = useState<ChainId | 'All'>('All')
 
@@ -90,52 +77,6 @@ export const AssetSearch: FC<AssetSearchProps> = ({
 
   const handleClick = onClick ?? defaultClickHandler
 
-  const marketData = useAppSelector(selectMarketDataSortedByMarketCap)
-  const sortedAssets = useMemo(() => {
-    const selectAssetFiatBalance = (asset: Asset) =>
-      bnOrZero(
-        accountId
-          ? portfolioFiatBalancesByAccount[accountId][asset.assetId]
-          : portfolioFiatBalances[asset.assetId],
-      ).toNumber()
-    const selectAssetMarketCap = (asset: Asset) =>
-      bnOrZero(marketData[asset.assetId]?.marketCap).toNumber()
-    const selectAssetVolume = (asset: Asset) =>
-      bnOrZero(marketData[asset.assetId]?.volume).toNumber()
-    const selectAssetChangePercent24Hr = (asset: Asset) =>
-      bnOrZero(marketData[asset.assetId]?.changePercent24Hr).toNumber()
-    const selectAssetName = (asset: Asset) => asset.name
-
-    switch (sortOrder) {
-      case AssetSearchOrder.MarketCap:
-        return orderBy(filteredAssets, [selectAssetMarketCap, selectAssetName], ['desc', 'asc'])
-      case AssetSearchOrder.Volume:
-        return orderBy(filteredAssets, [selectAssetVolume, selectAssetName], ['desc', 'asc'])
-      case AssetSearchOrder.ChangePercent24h:
-        return orderBy(
-          filteredAssets,
-          [selectAssetChangePercent24Hr, selectAssetName],
-          ['desc', 'asc'],
-        )
-      case AssetSearchOrder.UserBalance:
-        return orderBy(filteredAssets, [selectAssetFiatBalance, selectAssetName], ['desc', 'asc'])
-      case AssetSearchOrder.UserBalanceMarketCap:
-        return orderBy(
-          filteredAssets,
-          [selectAssetFiatBalance, selectAssetMarketCap, selectAssetName],
-          ['desc', 'desc', 'asc'],
-        )
-      default:
-        return filteredAssets
-    }
-  }, [
-    accountId,
-    filteredAssets,
-    marketData,
-    portfolioFiatBalances,
-    portfolioFiatBalancesByAccount,
-    sortOrder,
-  ])
   const [searchTermAssets, setSearchTermAssets] = useState<Asset[]>([])
   const { register, watch } = useForm<{ search: string }>({
     mode: 'onChange',
@@ -148,16 +89,16 @@ export const AssetSearch: FC<AssetSearchProps> = ({
   const searching = useMemo(() => searchString.length > 0, [searchString])
 
   useEffect(() => {
-    if (sortedAssets) {
+    if (filteredAssets) {
       setSearchTermAssets(
-        searching ? filterAssetsBySearchTerm(searchString, sortedAssets) : sortedAssets,
+        searching ? filterAssetsBySearchTerm(searchString, filteredAssets) : filteredAssets,
       )
     } else {
       moduleLogger.error('sortedAssets not defined')
     }
-  }, [searchString, searching, sortedAssets])
+  }, [searchString, searching, filteredAssets])
 
-  const listAssets = searching ? searchTermAssets : sortedAssets
+  const listAssets = searching ? searchTermAssets : filteredAssets
 
   /**
    * display a list of chain icon filters, based on a unique list of chain ids,
