+++ src/components/SelectAssets/SelectAssets.tsx
@@ -4,7 +4,7 @@ import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { AssetSearch } from 'components/AssetSearch/AssetSearch'
 import { SlideTransition } from 'components/SlideTransition'
-import { selectAssetsByMarketCap } from 'state/slices/selectors'
+import { selectSortedAssets } from 'state/slices/selectors'
 
 export type Asset = {}
 
@@ -15,7 +15,7 @@ type SelectAssetsProps = {
 
 export const SelectAssets = ({ onClick, onBack: handleBack }: SelectAssetsProps) => {
   const translate = useTranslate()
-  const assetsByMarketCap = useSelector(selectAssetsByMarketCap)
+  const sortedAssets = useSelector(selectSortedAssets)
   return (
     <SlideTransition>
       <Stack direction='row' width='full' alignItems='center' px={4}>
@@ -34,7 +34,7 @@ export const SelectAssets = ({ onClick, onBack: handleBack }: SelectAssetsProps)
         <ModalCloseButton position='static' />
       </Stack>
       <ModalBody height='600px' px={2} display='flex' flexDir='column'>
-        <AssetSearch onClick={onClick} assets={assetsByMarketCap} />
+        <AssetSearch onClick={onClick} assets={sortedAssets} />
       </ModalBody>
     </SlideTransition>
   )
